+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* PreferLocations is pure function now
+
 ## v3.74.2
 * Added description to scan errors with use query service client scanner